app/requirements.txt
Django==3.2.3
asgiref==3.3.4
certifi==2021.5.30
chardet===4.0.0
dj-config-url==0.1.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
html2text==2020.1.16
idna==2.10
lxml==4.5.0
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.5
pycurl==7.43.0.5
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
setuptools==40.6.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wikipedia==1.4.0
pytrends==4.7.3
beautifulsoup4==4.9.3
urlextract==1.3.0
wptools==0.4.17
scikit-learn==0.24.2







